Update strings.js
Changed buf2binstring() to use fast method by chunking large strings. I've tested this code in my project in the major browsers, but not in Pako and node. This is a few times faster in my tests with MB strings, except in Chrome where it is >8x faster (as string begins exceeding ~400kb). (Disclaimer: I'm fairly new to js and github)