Introducing late Uint8Array initialization
This is the fix for the latency issue when the initialization code is executed once it is loaded into runtime and causes an increase in parsing & execution time regardless of its usage. In particular, Uint8Array buffer initialization either triggers GC sweep or blocks a network thread in older browsers (ArrayBuffers used to be allocated from the network heap).